Make Services accept/reject all in selection.
Replaced the addition, deletion and substitution regular expression
searches with substitution. This way, it is possible to daisy chain the
outputs of the substitutions. This means that it's not longer necessary
to selection _only_ the critic markup section; it's possible to select
any text, as long as it contains the markup to accept/reject.

This means that it's possible to select a section with multiple edits,
even the whole document and "accept all" or "reject all".